config: Set screenshot keychord timeout to 0 ms
There's no point in having a negligible debounce time of 50 ms for the
screenshot keychord. Physical bouncing should be debounced by the
hardware or kernel driver (as is the case on Qualcomm PMICs), and
accidental pocket presses will not be mitigated by a 50 ms timeout.
Let's just set it to 0 ms to remove the delay altogether and reduce
screenshot latency.

Change-Id: I76cb8d606c81ced34a35c82f0cb1c4e58f803c06